+++ test/services/forest_liana/resource_updater_test.rb
@@ -12,38 +12,57 @@ module ForestLiana
     ForestLiana.apimap << collection
     ForestLiana.models << SerializeField
 
-    test 'SerializeField with attribute null' do
-      params = ActionController::Parameters.new(
-        id: 1,
-        data: { id: 1, type: "serialize_field", attributes: { } }
-      )
-      updater = ResourceUpdater.new(SerializeField, params)
-      updater.perform
-      assert updater.record.valid?
-    end
-
-    test 'SerializeField with bad format attribute' do
-      params = ActionController::Parameters.new(
-        id: 1,
-        data: { id: 1, type: "serialize_field", attributes: { field: "Lucas" } }
-      )
-      updater = ResourceUpdater.new(SerializeField, params)
-      updater.perform
-      assert updater.record.valid?
-    end
-
-    test 'SerializeField with right attribute' do
+    test 'Update a record on a "serialize" attribute with a missing value' do
       params = ActionController::Parameters.new(
         id: 1,
         data: {
-            id: 1,
-            type: "serialize_field",
-            attributes: { field: "[\"test\", \"test\"]" }
-          }
+          id: 1,
+          type: "serialize_field",
+          attributes: {}
+        }
       )
       updater = ResourceUpdater.new(SerializeField, params)
       updater.perform
+
       assert updater.record.valid?
+      assert updater.record.field == []
+    end
+
+    test 'Update a record on a "serialize" attribute with a bad format value' do
+      params = ActionController::Parameters.new(
+        id: 1,
+        data: {
+          id: 1,
+          type: "serialize_field",
+          attributes: {
+            field: "Lucas"
+          }
+        }
+      )
+      updater = ResourceUpdater.new(SerializeField, params)
+      updater.perform
+
+      assert updater.record.valid?
+      assert updater.record.field == "value 1"
+      assert updater.errors[0][:detail] == "Bad format for 'field' attribute value."
+    end
+
+    test 'Update a record on a "serialize" attribute with a well formated value' do
+      params = ActionController::Parameters.new(
+        id: 1,
+        data: {
+          id: 1,
+          type: "serialize_field",
+          attributes: {
+            field: "[\"test\", \"test\"]"
+          }
+        }
+      )
+      updater = ResourceUpdater.new(SerializeField, params)
+      updater.perform
+
+      assert updater.record.valid?
+      assert updater.record.field == ["test", "test"]
     end
   end
 end